refactor(ios, webview): extract methods.
